Address review feedback: split files, fix hardcoded string, update copy
- Extract CreateAccountDialogView into its own file (CreateAccountDialogView.tsx)
  to bring CreateAccountDialog.tsx under the 240-line split threshold
- Split CreateAccountDialog.spec.tsx (475 lines) into CreateAccountDialog-tests/
  with view.spec.tsx and integration.spec.tsx
- Add typePlaceholder key to copy.ts and reference it from the view instead of
  the hardcoded "Select a type…" string
- Re-export CreateAccountDialogView and its props type from CreateAccountDialog.tsx
  to preserve existing import paths used by the spec and stories files
